Moved model to different folder
