+++ svc-bip-api/src/integrationTest/java/gov/va/vro/bip/RabbitMqIntegrationTest.java
@@ -12,8 +12,8 @@ import gov.va.vro.bip.model.RequestForUpdateClaimStatus;
 import gov.va.vro.bip.model.UpdateContention;
 import gov.va.vro.bip.model.UpdateContentionModel;
 import gov.va.vro.bip.model.UpdateContentionReq;
-import gov.va.vro.bip.modelv2.contentions.GetClaimContentionsRequest;
-import gov.va.vro.bip.modelv2.contentions.GetClaimContentionsResponse;
+import gov.va.vro.bip.model.contentions.GetClaimContentionsRequest;
+import gov.va.vro.bip.model.contentions.GetClaimContentionsResponse;
 import gov.va.vro.bip.service.BipApiService;
 import gov.va.vro.bip.service.RabbitMqController;
 import lombok.SneakyThrows;
