+++ svc-bip-api/src/main/java/gov/va/vro/bip/model/BipPayloadResponse.java
@@ -1,10 +1,8 @@
-package gov.va.vro.bip.modelv2.contentions;
+package gov.va.vro.bip.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import gov.va.vro.bip.model.ClaimContention;
-import gov.va.vro.bip.modelv2.BipPayloadResponse;
 import lombok.Getter;
 import lombok.experimental.SuperBuilder;
 import lombok.extern.jackson.Jacksonized;
@@ -16,7 +14,13 @@ import java.util.List;
 @SuperBuilder(toBuilder = true)
 @JsonIgnoreProperties(ignoreUnknown = true)
 @JsonInclude(JsonInclude.Include.NON_EMPTY)
-public class GetClaimContentionsResponse extends BipPayloadResponse {
-  @JsonProperty("contentions")
-  private final List<ClaimContention> contentions;
+public class BipPayloadResponse {
+  @JsonProperty("statusCode")
+  private final int statusCode;
+
+  @JsonProperty("statusMessage")
+  private final String statusMessage;
+
+  @JsonProperty("messages")
+  private final List<BipMessage> messages;
 }
