+++ svc-bip-api/src/main/java/gov/va/vro/bip/model/contentions/GetClaimContentionsResponse.java
@@ -1,8 +1,10 @@
-package gov.va.vro.bip.modelv2;
+package gov.va.vro.bip.model.contentions;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
+import gov.va.vro.bip.model.BipPayloadResponse;
+import gov.va.vro.bip.model.ClaimContention;
 import lombok.Getter;
 import lombok.experimental.SuperBuilder;
 import lombok.extern.jackson.Jacksonized;
@@ -14,13 +16,7 @@ import java.util.List;
 @SuperBuilder(toBuilder = true)
 @JsonIgnoreProperties(ignoreUnknown = true)
 @JsonInclude(JsonInclude.Include.NON_EMPTY)
-public class BipPayloadResponse {
-  @JsonProperty("statusCode")
-  private final int statusCode;
-
-  @JsonProperty("statusMessage")
-  private final String statusMessage;
-
-  @JsonProperty("messages")
-  private final List<BipMessage> messages;
+public class GetClaimContentionsResponse extends BipPayloadResponse {
+  @JsonProperty("contentions")
+  private final List<ClaimContention> contentions;
 }
