+++ svc-bip-api/src/main/java/gov/va/vro/bip/service/RabbitMqController.java
@@ -4,8 +4,8 @@ import gov.va.vro.bip.model.BipClaimResp;
 import gov.va.vro.bip.model.BipUpdateClaimResp;
 import gov.va.vro.bip.model.RequestForUpdateClaimStatus;
 import gov.va.vro.bip.model.UpdateContentionModel;
-import gov.va.vro.bip.modelv2.contentions.GetClaimContentionsRequest;
-import gov.va.vro.bip.modelv2.contentions.GetClaimContentionsResponse;
+import gov.va.vro.bip.model.contentions.GetClaimContentionsRequest;
+import gov.va.vro.bip.model.contentions.GetClaimContentionsResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.amqp.rabbit.annotation.RabbitListener;
